# src/main/resources/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "接口说明文档",
    "version" : "v1",
    "title" : "接口说明文档"
  },
  "host" : "localhost:8080",
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/provider/approve/{providernum}" : {
      "put" : {
        "summary" : "认证供应商",
        "description" : "",
        "operationId" : "providerApprove",
        "parameters" : [ {
          "name" : "providernum",
          "in" : "path",
          "description" : "供应商账号",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "忽略供应商认证请求",
        "description" : "",
        "operationId" : "providerIgnore",
        "parameters" : [ {
          "name" : "providernum",
          "in" : "path",
          "description" : "供应商账号",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
    "/provider/message" : {
      "get" : {
        "summary" : "供应商信息跳转",
        "description" : "",
        "operationId" : "providerMsgDetail",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Model"
          }
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/HttpServletRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/provider/message/{providernum}" : {
      "put" : {
        "summary" : "更新供应商信息",
        "description" : "",
        "operationId" : "providerMsg",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "提交的供应商信息",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ProviderMsg"
          }
        }, {
          "name" : "providernum",
          "in" : "path",
          "description" : "供应商账号",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/HttpServletRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
...
    "/provider/reject" : {
      "get" : {
        "summary" : "当前供应商的问题订单",
        "description" : "",
        "operationId" : "providerReject",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/HttpServletRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/RestMessage"
            }
          }
        }
      }
    },
    "/provider/send/{ordernumber}" : {
      "put" : {
        "summary" : "确认发货",
        "description" : "",
        "operationId" : "providerSend",
        "parameters" : [ {
          "name" : "ordernumber",
          "in" : "path",
          "description" : "确认发货",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/HttpServletRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "BufferedReader" : {
      "type" : "object"
    },
    "Cookie" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        },
        "comment" : {
          "type" : "string"
        },
        "domain" : {
          "type" : "string"
        },
        "maxAge" : {
          "type" : "integer",
          "format" : "int32"
        },
        "path" : {
          "type" : "string"
        },
        "secure" : {
          "type" : "boolean",
          "default" : false
        },
        "version" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "Enumeration" : {
      "type" : "object"
    },
    "HttpServletRequest" : {
      "type" : "object",
      "properties" : {
        "method" : {
          "type" : "string"
        },
        "requestedSessionIdValid" : {
          "type" : "boolean",
          "default" : false
        },
        "pathTranslated" : {
          "type" : "string"
        },
        "requestedSessionIdFromCookie" : {
          "type" : "boolean",
          "default" : false
        },
        "requestedSessionIdFromURL" : {
          "type" : "boolean",
          "default" : false
        },
        "requestedSessionIdFromUrl" : {
          "type" : "boolean",
          "default" : false
        },
        "requestedSessionId" : {
          "type" : "string"
        },
        "session" : {
          "$ref" : "#/definitions/HttpSession"
        },
        "userPrincipal" : {
          "$ref" : "#/definitions/Principal"
        },
        "contextPath" : {
          "type" : "string"
        },
        "requestURL" : {
          "$ref" : "#/definitions/StringBuffer"
        },
        "headerNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "remoteUser" : {
          "type" : "string"
        },
        "authType" : {
          "type" : "string"
        },
        "pathInfo" : {
          "type" : "string"
        },
        "servletPath" : {
          "type" : "string"
        },
        "requestURI" : {
          "type" : "string"
        },
        "queryString" : {
          "type" : "string"
        },
        "cookies" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Cookie"
          }
        },
        "scheme" : {
          "type" : "string"
        },
        "inputStream" : {
          "$ref" : "#/definitions/ServletInputStream"
        },
        "protocol" : {
          "type" : "string"
        },
        "reader" : {
          "$ref" : "#/definitions/BufferedReader"
        },
        "contentLength" : {
          "type" : "integer",
          "format" : "int32"
        },
        "contentType" : {
          "type" : "string"
        },
        "locale" : {
          "$ref" : "#/definitions/Locale"
        },
        "attributeNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "characterEncoding" : {
          "type" : "string"
        },
        "parameterMap" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        },
        "parameterNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "localName" : {
          "type" : "string"
        },
        "remoteHost" : {
          "type" : "string"
        },
        "remotePort" : {
          "type" : "integer",
          "format" : "int32"
        },
        "serverName" : {
          "type" : "string"
        },
        "localPort" : {
          "type" : "integer",
          "format" : "int32"
        },
        "locales" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "remoteAddr" : {
          "type" : "string"
        },
        "secure" : {
          "type" : "boolean",
          "default" : false
        },
        "localAddr" : {
          "type" : "string"
        },
        "serverPort" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "HttpSession" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "creationTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "attributeNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "servletContext" : {
          "$ref" : "#/definitions/ServletContext"
        },
        "maxInactiveInterval" : {
          "type" : "integer",
          "format" : "int32"
        },
        "sessionContext" : {
          "$ref" : "#/definitions/HttpSessionContext"
        },
        "lastAccessedTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "valueNames" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "new" : {
          "type" : "boolean",
          "default" : false
        }
      }
    },
    "HttpSessionContext" : {
      "type" : "object",
      "properties" : {
        "ids" : {
          "$ref" : "#/definitions/Enumeration"
        }
      }
    },
    "Locale" : {
      "type" : "object",
      "properties" : {
        "language" : {
          "type" : "string"
        },
        "country" : {
          "type" : "string"
        },
        "displayCountry" : {
          "type" : "string"
        },
        "displayLanguage" : {
          "type" : "string"
        },
        "displayName" : {
          "type" : "string"
        },
        "displayScript" : {
          "type" : "string"
        },
        "displayVariant" : {
          "type" : "string"
        },
        "extensionKeys" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        },
        "iso3Country" : {
          "type" : "string"
        },
        "iso3Language" : {
          "type" : "string"
        },
        "script" : {
          "type" : "string"
        },
        "unicodeLocaleAttributes" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        },
        "unicodeLocaleKeys" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        },
        "variant" : {
          "type" : "string"
        }
      }
    },
    "Model" : {
      "type" : "object"
    },
...
    "Principal" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        }
      }
    },
    "ProviderMsg" : {
      "type" : "object",
      "properties" : {
        "providernum" : {
          "type" : "string"
        },
        "shopname" : {
          "type" : "string"
        },
        "connect" : {
          "type" : "string"
        },
        "shopmsg" : {
          "type" : "string"
        },
        "location" : {
          "type" : "string"
        },
        "license" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string"
        }
      }
    },
    "RestMessage" : {
      "type" : "object",
      "properties" : {
        "count" : {
          "type" : "string",
          "description" : "数据数量"
        },
        "data" : {
          "type" : "object",
          "description" : "返回数据"
        },
        "code" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "返回编号"
        },
        "msg" : {
          "type" : "string",
          "description" : "返回信息"
        }
      },
      "description" : "返回信息"
    },
    "ServletContext" : {
      "type" : "object",
      "properties" : {
        "attributeNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "initParameterNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "servletContextName" : {
          "type" : "string"
        },
        "majorVersion" : {
          "type" : "integer",
          "format" : "int32"
        },
        "minorVersion" : {
          "type" : "integer",
          "format" : "int32"
        },
        "contextPath" : {
          "type" : "string"
        },
        "servlets" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "servletNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "serverInfo" : {
          "type" : "string"
        }
      }
    },
    "ServletInputStream" : {
      "type" : "object"
    },
    "StringBuffer" : {
      "type" : "object"
    },
...
  }
}